divided the js file into modules
